test: extract outbound session route coverage
(cherry picked from commit d537904abb52c4e99ca6697bcf3a8d3c6d6fd136)
