+++ src/infra/outbound/outbound.test.ts
@@ -24,7 +24,6 @@ import {
   formatGatewaySummary,
   formatOutboundDeliverySummary,
 } from "./format.js";
-import { resolveOutboundSessionRoute } from "./outbound-session.js";
 import {
   formatOutboundPayloadLog,
   normalizeOutboundPayloads,
@@ -794,251 +793,6 @@ describe("formatGatewaySummary", () => {
     for (const testCase of cases) {
       expect(formatGatewaySummary(testCase.input), testCase.name).toBe(testCase.expected);
     }
-  });
-});
-
-describe("resolveOutboundSessionRoute", () => {
-  const baseConfig = {} as RemoteClawConfig;
-
-  it("resolves provider-specific session routes", async () => {
-    const perChannelPeerCfg = { session: { dmScope: "per-channel-peer" } } as RemoteClawConfig;
-    const identityLinksCfg = {
-      session: {
-        dmScope: "per-peer",
-        identityLinks: {
-          alice: ["discord:123"],
-        },
-      },
-    } as RemoteClawConfig;
-    const slackMpimCfg = {
-      channels: {
-        slack: {
-          dm: {
-            groupChannels: ["G123"],
-          },
-        },
-      },
-    } as RemoteClawConfig;
-    const cases: Array<{
-      name: string;
-      cfg: RemoteClawConfig;
-      channel: string;
-      target: string;
-      replyToId?: string;
-      threadId?: string;
-      expected: {
-        sessionKey: string;
-        from?: string;
-        to?: string;
-        threadId?: string | number;
-        chatType?: "direct" | "group" | "channel";
-      };
-    }> = [
-      {
-        name: "Slack thread",
-        cfg: baseConfig,
-        channel: "slack",
-        target: "channel:C123",
-        replyToId: "456",
-        expected: {
-          sessionKey: "agent:main:slack:channel:c123:thread:456",
-          from: "slack:channel:C123",
-          to: "channel:C123",
-          threadId: "456",
-        },
-      },
-      {
-        name: "Telegram topic group",
-        cfg: baseConfig,
-        channel: "telegram",
-        target: "-100123456:topic:42",
-        expected: {
-          sessionKey: "agent:main:telegram:group:-100123456:topic:42",
-          from: "telegram:group:-100123456:topic:42",
-          to: "telegram:-100123456",
-          threadId: 42,
-        },
-      },
-      {
-        name: "Telegram unresolved username DM",
-        cfg: perChannelPeerCfg,
-        channel: "telegram",
-        target: "@alice",
-        expected: {
-          sessionKey: "agent:main:telegram:direct:@alice",
-          chatType: "direct",
-        },
-      },
-      {
-        name: "Telegram DM scoped threadId fallback",
-        cfg: perChannelPeerCfg,
-        channel: "telegram",
-        target: "12345",
-        threadId: "12345:99",
-        expected: {
-          sessionKey: "agent:main:telegram:direct:12345",
-          from: "telegram:12345",
-          to: "telegram:12345",
-          threadId: 99,
-          chatType: "direct",
-        },
-      },
-      {
-        name: "identity-links per-peer",
-        cfg: identityLinksCfg,
-        channel: "discord",
-        target: "user:123",
-        expected: {
-          sessionKey: "agent:main:direct:alice",
-        },
-      },
-      {
-        name: "BlueBubbles chat_* prefix stripping",
-        cfg: baseConfig,
-        channel: "bluebubbles",
-        target: "chat_guid:ABC123",
-        expected: {
-          sessionKey: "agent:main:bluebubbles:group:abc123",
-          from: "group:ABC123",
-        },
-      },
-      {
-        name: "Zalo Personal DM target",
-        cfg: perChannelPeerCfg,
-        channel: "zalouser",
-        target: "123456",
-        expected: {
-          sessionKey: "agent:main:zalouser:direct:123456",
-          chatType: "direct",
-        },
-      },
-      {
-        name: "Slack mpim allowlist -> group key",
-        cfg: slackMpimCfg,
-        channel: "slack",
-        target: "channel:G123",
-        expected: {
-          sessionKey: "agent:main:slack:group:g123",
-          from: "slack:group:G123",
-        },
-      },
-      {
-        name: "Feishu explicit group prefix keeps group routing",
-        cfg: baseConfig,
-        channel: "feishu",
-        target: "group:oc_group_chat",
-        expected: {
-          sessionKey: "agent:main:feishu:group:oc_group_chat",
-          from: "feishu:group:oc_group_chat",
-          to: "oc_group_chat",
-          chatType: "group",
-        },
-      },
-      {
-        name: "Feishu explicit dm prefix keeps direct routing",
-        cfg: perChannelPeerCfg,
-        channel: "feishu",
-        target: "dm:oc_dm_chat",
-        expected: {
-          sessionKey: "agent:main:feishu:direct:oc_dm_chat",
-          from: "feishu:oc_dm_chat",
-          to: "oc_dm_chat",
-          chatType: "direct",
-        },
-      },
-      {
-        name: "Feishu bare oc_ target defaults to direct routing",
-        cfg: perChannelPeerCfg,
-        channel: "feishu",
-        target: "oc_ambiguous_chat",
-        expected: {
-          sessionKey: "agent:main:feishu:direct:oc_ambiguous_chat",
-          from: "feishu:oc_ambiguous_chat",
-          to: "oc_ambiguous_chat",
-          chatType: "direct",
-        },
-      },
-      {
-        name: "Slack user DM target",
-        cfg: perChannelPeerCfg,
-        channel: "slack",
-        target: "user:U12345ABC",
-        expected: {
-          sessionKey: "agent:main:slack:direct:u12345abc",
-          from: "slack:U12345ABC",
-          to: "user:U12345ABC",
-          chatType: "direct",
-        },
-      },
-      {
-        name: "Slack channel target without thread",
-        cfg: baseConfig,
-        channel: "slack",
-        target: "channel:C999XYZ",
-        expected: {
-          sessionKey: "agent:main:slack:channel:c999xyz",
-          from: "slack:channel:C999XYZ",
-          to: "channel:C999XYZ",
-          chatType: "channel",
-        },
-      },
-    ];
-
-    for (const testCase of cases) {
-      const route = await resolveOutboundSessionRoute({
-        cfg: testCase.cfg,
-        channel: testCase.channel,
-        agentId: "main",
-        target: testCase.target,
-        replyToId: testCase.replyToId,
-        threadId: testCase.threadId,
-      });
-      expect(route?.sessionKey, testCase.name).toBe(testCase.expected.sessionKey);
-      if (testCase.expected.from !== undefined) {
-        expect(route?.from, testCase.name).toBe(testCase.expected.from);
-      }
-      if (testCase.expected.to !== undefined) {
-        expect(route?.to, testCase.name).toBe(testCase.expected.to);
-      }
-      if (testCase.expected.threadId !== undefined) {
-        expect(route?.threadId, testCase.name).toBe(testCase.expected.threadId);
-      }
-      if (testCase.expected.chatType !== undefined) {
-        expect(route?.chatType, testCase.name).toBe(testCase.expected.chatType);
-      }
-    }
-  });
-
-  it("uses resolved Discord user targets to route bare numeric ids as DMs", async () => {
-    const route = await resolveOutboundSessionRoute({
-      cfg: { session: { dmScope: "per-channel-peer" } } as RemoteClawConfig,
-      channel: "discord",
-      agentId: "main",
-      target: "123",
-      resolvedTarget: {
-        to: "user:123",
-        kind: "user",
-        source: "directory",
-      },
-    });
-
-    expect(route).toMatchObject({
-      sessionKey: "agent:main:discord:direct:123",
-      from: "discord:123",
-      to: "user:123",
-      chatType: "direct",
-    });
-  });
-
-  it("rejects bare numeric Discord targets when the caller has no kind hint", async () => {
-    await expect(
-      resolveOutboundSessionRoute({
-        cfg: { session: { dmScope: "per-channel-peer" } } as RemoteClawConfig,
-        channel: "discord",
-        agentId: "main",
-        target: "123",
-      }),
-    ).rejects.toThrow(/Ambiguous Discord recipient/);
   });
 });
